refresh the project structure
